Move sub-entry into component directory.
