+++ src/app/components/App.tsx
@@ -125,7 +125,7 @@ const UniRefEntryPage = lazy(
 const UniParcSubEntryPage = lazy(
   () =>
     import(
-      /* webpackChunkName: "uniparc-entry" */ '../../uniparc/components/entry/sub-entry/SubEntry'
+      /* webpackChunkName: "uniparc-entry" */ '../../uniparc/components/sub-entry/SubEntry'
     )
 );
 const UniParcEntryPage = lazy(
